cushion {plotflow}	R Documentation

Determine Max Value + n Extra

Description

Determine max value of a vector plus an additional pproportion.

Usage

cushion(x, pad = 0.05)

Arguments

x
A vector.

pad
A proportion extra to add.

Value

Returns the max value of a vector + n extra.

Note

If x is a factor or character vector counts for each category are used as max(n).

Examples

cushion(as.factor(mtcars$cyl))
cushion(mtcars$cyl)
cushion(mtcars$cyl, .5)

## Not run: 
ggplot(reorder_by(cyl, ~-cyl , mtcars, length), aes(x=as.factor(cyl))) + 
    geom_bar()  + 
    theme_apa() + 
    y0(cushion(as.factor(mtcars$cyl))) +
    xlab("Cylinders") + 
    ylab("Total")

## End(Not run)

[Package plotflow version 0.1.0 Index]