plot2drop {plotflow}	R Documentation

Plot to Drop Box

Description

A wrapper to plotting device for easy plotting to Drop Box.

Usage

plot2drop(name, type = "png", width = 400, height = 400, open = TRUE,
  loc = getOption("dropbox_path"), key = getOption("dropbox_key"), ...)

Arguments

name
The name of the plot (no file extension needed).

type
The type of plotting device (png, pdf, etc).

width
Width of plot.

height
Height of plot.

open
logical. If TRUE provides the code to open the plot in default browser.

loc
The path to the location of the dropbox public folder.

key
The dropbox key.

...
Other arguments passed to the plotting device in type.

Details

Plots to Drop Box and returns the URL to the plot (and optionally opens the plot). Code attemoted to be copied to the clipboard.

Examples

## Not run: 
plot2drop("dfg")
plot(1:10)
dev.off()

## End(Not run)

[Package plotflow version 0.1.0 Index]