theme_apa {plotflow}	R Documentation

ggplot2 Theme for APA Publications

Description

A ggplot2 theme with no background and Times New Roman font.

Usage

theme_apa(plot.box = FALSE)

Arguments

plot.box
logical. If TRUE a full box surrounds the plot area. If FALSE only the x and y axis are shown.

See Also

theme

Examples

## Not run: 
ggplot(reorder_by(cyl, ~-cyl , mtcars, length), aes(x=as.factor(cyl))) + 
    geom_bar()  + 
    theme_apa() + 
    y0(cushion(as.factor(mtcars$cyl))) +
    xlab("Cylinders") + 
    ylab("Total")

ggplot(reorder_by(cyl, ~-cyl , mtcars, length), aes(x=as.factor(cyl))) + 
    geom_bar()  + 
    theme_apa(plot.box=T) + 
    y0(cushion(as.factor(mtcars$cyl))) +
    xlab("Cylinders") + 
    ylab("Total")

ggplot(reorder_by(cyl, ~-cyl , mtcars, length), aes(x=as.factor(cyl))) + 
    geom_bar()  + 
    theme_basic() +
    theme_apa() + 
    y0(cushion(as.factor(mtcars$cyl))) +
    xlab("Cylinders") + 
    ylab("Total")

## End(Not run)

[Package plotflow version 0.1.0 Index]